Fixing side effects of moving LinkButton
